<update> Updated functions & function calls.
Updated the Equals () function to contain the IF statement for a,b, & operand. clearCurrentArray() was not working as intended so added in the clearedArray variable. 
The number buttons eventlistener now also calls the clearCurrentArray(), this needs to be changed so that this is only called once for the previous array, i.e if currentinteger === "str" then clear the current array, will take some testing to get it right. 

After the above it should be a case of ensuring that only the dedicated options can be chosen and the screen updates with the new value of a(). the let total = null; has been removed due to not needed. This does throw a spanner in the works as it means that you can only do one operation at a time, so it cant be (total +(4+5) = new total etc. The project doesn't require this but would be good to figure it out.